Setup dry-system with better configuration
Upgrade dry-system to master.
Split Web and Loot into two distinct folders.
Setup notifications and logger.
