+++ apps/web/controllers/characters/show.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+module Web
+  module Controllers
+    module Characters
+      class Show
+        include Hanami::Action
+        include Assistant::Import['repos.character_repo']
+
+        format 'json'
+
+        configuration do
+          default_response_format :json
+        end
+
+        def call(params)
+          character = character_repo.by_id(params[:id])
+
+          character_serializer = Serializers::CharacterSerializer.new
+
+          self.format = :json
+          self.body = character_serializer.serialize(character).to_json
+          self.status = 200
+        end
+      end
+    end
+  end
+end
