+++ bin/console
@@ -3,15 +3,9 @@
 
 require_relative '../config/application'
 
-Loot::Application.finalize!
+Assistant::finalize_application!
 
 require 'pry'
 require 'pry-byebug'
-
-def scenario
-  data = JSON.parse(File.read('/home/alex/Downloads/message.txt'))
-
-  Loot::Attribution::Imports::RcLootCouncilImporter.new.from_json(data)
-end
 
 Pry.start
