+++ lib/loot/attribution/imports/rc_loot_council_resource_extractor.rb
