+++ lib/loot/repos/character_repo.rb
@@ -5,7 +5,7 @@ module Loot
     class CharacterRepo < ROM::Repository[:characters]
       struct_namespace ::Loot::Entities
 
-      include Import['container']
+      include Assistant::Import['container']
 
       commands :create,
                use: :timestamps,
