--- lib/loot/web/controllers/root/index.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module Loot
-  module Web
-    module Controllers
-      module Root
-        class Index
-          include Hanami::Action
-          format 'json'
-
-          def call(_params)
-            self.body = {}.to_json
-            self.status = 200
-          end
-        end
-      end
-    end
-  end
-end
