+++ system/boot/persistence.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-Loot::Application.boot(:persistence) do |app|
+Assistant::Application.boot(:persistence) do |app|
   start do
     config = app['db.config']
     config.auto_registration("#{app.root}/lib/loot")
